refactor(enduser-ui-fe): extract ManagerNexus components
* Extracted header, footer, HUD grid, empty state, and spec panel into `features/manager/components/nexus/`.
* Extracted `EthicsAuditPanel` and fully encapsulated the `handleViewDiff` method to reduce main file complexity.
* Maintained existing typing and styling conventions.
* Reduced bundle size of `ManagerNexus.tsx` by isolating `framer-motion` and `react-markdown` to the spec panel.
